Removed any "C:\Users\..." hard-coded values from the code. Meshes are stored using the unique_ptr instead of direct instances in the MeshFactory. Fonts are now handled more efficiently in the FontFactory class. Replaced "typedef" in Utility module with "using". Windows-related header files are no longer included in the code where it is not necessary.

+++ Audio/FMOD_AudioContext.cpp
@@ -30,7 +30,7 @@ void Audio::Fmod::FmodAudioContext::StopAudio(AudioObject& audioObject)
 	}
 }
 
-void Audio::Fmod::FmodAudioContext::GenerateSamples(UINT8* stream, int streamLength)
+void Audio::Fmod::FmodAudioContext::GenerateSamples(int* stream, int streamLength)
 {
 }
